+++ miasm/jitter/arch/JitCore_ppc32.c
@@ -54,7 +54,7 @@ cpu_set_gpreg(JitCpu *self, PyObject *args) {
 	for (i=0; i < sizeof(gpreg_dict)/sizeof(reg_dict); i++){
 	    if (strcmp(d_key_name, gpreg_dict[i].name))
 		continue;
-	    *((uint32_t*)(((char*)(self->cpu)) + gpreg_dict[i].offset)) = val;
+	    *((uint32_t*)(((char*)(self->cpu)) + gpreg_dict[i].offset)) = (uint32_t)val;
 	    found = 1;
 	    break;
 	}
